Tpetra: Fix edge case in sort(AndMerge)CrsEntries
@trilinos/tpetra In Tpetra::Import_Util, sortCrsEntries and
sortAndMergeCrsEntries were computing the number of rows over which to
iterate as follows:

size_t NumRows = CRS_rowptr.dimension_0() - 1;

This is bad, because size_t is usually (though not always) unsigned.
If the matrix has zero rows and has not been initialized, it's
possible for CRS_rowptr (the row offsets array) to have length 0.  If
size_t is unsigned, this will result in a very large positive number,
the result of casting -1 to unsigned.

This commit fixes this issue in sortCrsEntries and
sortAndMergeCrsEntries.